Factor out ForestConnection
