+++ pyquil/api/_base_connection.py
@@ -19,6 +19,7 @@ from __future__ import print_function
 import os
 import re
 import time
+import warnings
 from json.decoder import JSONDecodeError
 
 import requests
@@ -26,7 +27,10 @@ from requests.adapters import HTTPAdapter
 from six import integer_types
 from urllib3 import Retry
 
+from pyquil import Program
+from pyquil.api.job import Job
 from pyquil.api.errors import error_mapping, UnknownApiError, TooManyQubitsError
+from pyquil.wavefunction import Wavefunction
 from ._config import PyquilConfig
 
 TYPE_EXPECTATION = "expectation"
@@ -178,3 +182,217 @@ def validate_run_items(run_items):
 
 def get_job_id(response):
     return response.json()['jobId']
+
+
+def run_and_measure_payload(quil_program, qubits, trials, random_seed):
+    """REST payload for :py:func:`run_and_measure`"""
+    if not quil_program:
+        raise ValueError("You have attempted to run an empty program."
+                         " Please provide gates or measure instructions to your program.")
+
+    if not isinstance(quil_program, Program):
+        raise TypeError("quil_program must be a Quil program object")
+    validate_run_items(qubits)
+    if not isinstance(trials, integer_types):
+        raise TypeError("trials must be an integer")
+
+    payload = {"type": TYPE_MULTISHOT_MEASURE,
+               "qubits": list(qubits),
+               "trials": trials,
+               'compiled-quil': quil_program.out()}
+
+    if random_seed is not None:
+        payload['rng-seed'] = random_seed
+
+    return payload
+
+
+def wavefunction_payload(quil_program, classical_addresses, random_seed):
+    """REST payload for :py:func:`wavefunction`"""
+    if not isinstance(quil_program, Program):
+        raise TypeError("quil_program must be a Quil program object")
+    validate_run_items(classical_addresses)
+
+    payload = {'type': TYPE_WAVEFUNCTION,
+               'addresses': list(classical_addresses),
+               'compiled-quil': quil_program.out()}
+
+    if random_seed is not None:
+        payload['rng-seed'] = random_seed
+
+    return payload
+
+
+def expectation_payload(prep_prog, operator_programs, random_seed):
+    """REST payload for :py:func:`expectation`"""
+    if operator_programs is None:
+        operator_programs = [Program()]
+
+    if not isinstance(prep_prog, Program):
+        raise TypeError("prep_prog variable must be a Quil program object")
+
+    payload = {'type': TYPE_EXPECTATION,
+               'state-preparation': prep_prog.out(),
+               'operators': [x.out() for x in operator_programs]}
+
+    if random_seed is not None:
+        payload['rng-seed'] = random_seed
+
+    return payload
+
+
+class ForestConnection:
+    def __init__(self, sync_endpoint=SYNC_ENDPOINT,
+                 async_endpoint=ASYNC_ENDPOINT, api_key=None, user_id=None,
+                 use_queue=False, ping_time=0.1, status_time=2):
+        """
+        Represents a connection to Forest containing methods to wrap all possible API endpoints.
+
+        Users should not use methods from this class directly.
+
+        :param sync_endpoint: The endpoint of the server for running small jobs
+        :param async_endpoint: The endpoint of the server for running large jobs
+        :param api_key: The key to the Forest API Gateway (default behavior is to read from
+            config file)
+        :param user_id: Your userid for Forest (default behavior is to read from config file)
+        :param bool use_queue: Disabling this parameter may improve performance for small,
+            quick programs. To support larger programs, set it to True. (default: False)
+            *_async methods will always use the queue; See https://go.rigetti.com/connections
+            for more information.
+        :param int ping_time: Time in seconds for how long to wait between polling the server
+            for updated status information on a job. Note that this parameter doesn't matter if
+            use_queue is False.
+        :param int status_time: Time in seconds for how long to wait between printing status
+            information. To disable printing of status entirely then set status_time to False.
+            Note that this parameter doesn't matter if use_queue is False.
+        """
+        self.async_endpoint = async_endpoint
+        self.sync_endpoint = sync_endpoint
+        self.session = get_session(api_key, user_id)
+
+        self.use_queue = use_queue
+        self.ping_time = ping_time
+        self.status_time = status_time
+
+    def run_and_measure(self, quil_program, qubits, trials, random_seed):
+        """
+        Run a Forest ``run_and_measure`` job.
+
+        Users should use :py:func:`WavefunctionSimulator.run_and_measure` instead of calling
+        this directly.
+        """
+        payload = run_and_measure_payload(quil_program, qubits, trials, random_seed)
+        if self.use_queue:
+            response = post_json(self.session, self.async_endpoint + "/job",
+                                 {"machine": "QVM", "program": payload})
+            job = self.wait_for_job(get_job_id(response))
+            return job.result()
+        else:
+            response = post_json(self.session, self.sync_endpoint + "/qvm", payload)
+            return response.json()
+
+    def run_and_measure_async(self, quil_program, qubits, trials, random_seed):
+        """
+        Run a Forest ``run_and_measure`` job asynchronously.
+
+        Users should use :py:func:`WavefunctionSimulator.run_and_measure_async` instead of calling
+        this directly.
+        """
+        payload = run_and_measure_payload(quil_program, qubits, trials, random_seed)
+        response = post_json(self.session, self.async_endpoint + "/job",
+                             {"machine": "QVM", "program": payload})
+        return get_job_id(response)
+
+    def wavefunction(self, quil_program, classical_addresses, random_seed):
+        """
+        Run a Forest ``wavefunction`` job.
+
+        Users should use :py:func:`WavefunctionSimulator.wavefunction` instead of calling
+        this directly.
+        """
+        if self.use_queue:
+            payload = wavefunction_payload(quil_program, classical_addresses, random_seed)
+            response = post_json(self.session, self.async_endpoint + "/job",
+                                 {"machine": "QVM", "program": payload})
+            job = self.wait_for_job(get_job_id(response))
+            return job.result()
+        else:
+            payload = wavefunction_payload(quil_program, classical_addresses, random_seed)
+            response = post_json(self.session, self.sync_endpoint + "/qvm", payload)
+            return Wavefunction.from_bit_packed_string(response.content, classical_addresses)
+
+    def wavefunction_async(self, quil_program, classical_addresses, random_seed):
+        """
+        Run a Forest ``wavefunction`` job asynchronously.
+
+        Users should use :py:func:`WavefunctionSimulator.wavefunction_async` instead of calling
+        this directly.
+        """
+        payload = wavefunction_payload(quil_program, classical_addresses, random_seed)
+        response = post_json(self.session, self.async_endpoint + "/job",
+                             {"machine": "QVM", "program": payload})
+        return get_job_id(response)
+
+    def expectation(self, prep_prog, operator_programs, random_seed):
+        """
+        Run a Forest ``expectation`` job.
+
+        Users should use :py:func:`WavefunctionSimulator.expectation` instead of calling
+        this directly.
+        """
+        if isinstance(operator_programs, Program):
+            warnings.warn("You have provided a Program rather than a list of Programs. The results "
+                          "from expectation will be line-wise expectation values of the "
+                          "operator_programs.", SyntaxWarning)
+        if self.use_queue:
+            payload = expectation_payload(prep_prog, operator_programs, random_seed)
+            response = post_json(self.session, self.async_endpoint + "/job",
+                                 {"machine": "QVM", "program": payload})
+            job = self.wait_for_job(get_job_id(response))
+            return job.result()
+        else:
+            payload = expectation_payload(prep_prog, operator_programs, random_seed)
+            response = post_json(self.session, self.sync_endpoint + "/qvm", payload)
+            return response.json()
+
+    def expectation_async(self, prep_prog, operator_programs, random_seed):
+        """
+        Run a Forest ``expectation`` job asynchronously.
+
+        Users should use :py:func:`WavefunctionSimulator.expectation_async` instead of calling
+        this directly.
+        """
+        payload = expectation_payload(prep_prog, operator_programs, random_seed)
+        response = post_json(self.session, self.async_endpoint + "/job",
+                             {"machine": "QVM", "program": payload})
+        return get_job_id(response)
+
+    def get_job(self, job_id):
+        """
+        Given a job id, return information about the status of the job
+
+        :param str job_id: job id
+        :return: Job object with the status and potentially results of the job
+        :rtype: Job
+        """
+        response = get_json(self.session, self.async_endpoint + "/job/" + job_id)
+        return Job(response.json(), 'QVM')
+
+    def wait_for_job(self, job_id, ping_time=None, status_time=None):
+        """
+        Wait for the results of a job and periodically print status
+
+        :param job_id: Job id
+        :param ping_time: How often to poll the server.
+                          Defaults to the value specified in the constructor. (0.1 seconds)
+        :param status_time: How often to print status, set to False to never print status.
+                            Defaults to the value specified in the constructor (2 seconds)
+        :return: Completed Job
+        """
+
+        def get_job_fn():
+            return self.get_job(job_id)
+
+        return wait_for_job(get_job_fn,
+                            ping_time if ping_time else self.ping_time,
+                            status_time if status_time else self.status_time)
